[refactor] Move all commands working with installed systems to `quill installed`.
