+++ lib/quill/alias.js
@@ -21,9 +21,9 @@ quill.alias('uninstall', { resource: 'systems', command: 'uninstall' });
 //
 quill.alias('hostname',  { resource: 'keys',    command: 'hostname' });
 quill.alias('publish',   { resource: 'systems', command: 'publish' });
-quill.alias('installed', { resource: 'systems', command: 'installed' });
 quill.alias('unpublish', { resource: 'systems', command: 'unpublish' });
 quill.alias('bump',      { resource: 'systems', command: 'bump' });
 quill.alias('pack',      { resource: 'systems', command: 'pack' });
 quill.alias('list',      { resource: 'systems', command: 'list' });
 quill.alias('view',      { resource: 'systems', command: 'view' });
+quill.alias('latest', { resource: 'installed', command: 'latest' });
