+++ lib/quill/commands/systems.js
@@ -498,60 +498,3 @@ systems.bump.usage = [
   '',
   'quill systems bump [<release>]',
 ];
-
-//
-// ### function installed (callback)
-// #### @callback {function} Continuation to respond to.
-// Lists all systems installed on the current machine.
-//
-systems.installed = function (callback) {
-  composer.installed.list(function (err, list) {
-    if (err) {
-      return callback(err);
-    }
-
-    var names = Object.keys(list);
-
-    if (!names || !names.length) {
-      quill.log.warn('No systems installed');
-    }
-    else {
-      archy(composer.hierarchy(
-        quill.config.get('directories:install'),
-        names.reduce(function (systems, name) {
-          systems[name] = list[name].system;
-          return systems;
-        }, {})
-      ))
-      .split('\n')
-      .filter(Boolean)
-      .forEach(function (line) {
-        quill.log.data(line);
-      });
-    }
-
-    callback();
-  });
-};
-
-//
-// Usage for `quill systems installed`.
-//
-systems.installed.usage = [
-  'Lists all systems installed on the current machine.',
-  '',
-  'quill systems installed',
-  'quill installed'
-];
-
-systems.latest = function (name, callback) {
-  quill.log.info('Latest versions that satisfy all semver requirements.');
-  composer.maxSatisfying(name, function (err, latest) {
-    if (err) {
-      return callback(err);
-    }
-
-    quill.inspect.putObject(latest);
-    callback(null, latest);
-  });
-}
